[FEATURE] Introduce signal in loadBaseTca
To improve performance in backend context, final
TCA should be cached in contrast to only parts of it.

To be able to cache the complete TCA, we need to introduce a signal
after building the cached TCA from files.

#57862 describes further steps, which are not in scope of this change.

This is a completely non breaking change, it only
adds new API which extension authors can also
use to benefit from the resulting performance gain.

Resolves: #57863
Documentation: #57895
Releases: 6.2
Change-Id: I1174b68b59f2ebc5b8c0c227ac13102366c2136a
Reviewed-on: https://review.typo3.org/29433
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind