[BUGFIX] Make category TCA changes cacheable
Currently TCA changes of the category API are applied on every request
after the extTables files is included. This makes it impossible to
override the changes by third party extensions and also we do not
benefit from caching.

Instead we now use the introduced signal and apply the changes
right before caching.

With this change we also have a clear separation of
registering category field additions (in ext_localconf.php) and
applying the TCA changes while building the TCA cache.

Resolves: #57881
Releases: 6.2
Change-Id: I385c894fe13eb4cd4f8f2b4c82a9513b05308daa
Reviewed-on: https://review.typo3.org/29424
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring