[BUGFIX] Gracefully show missing files in IRRE relations
When the database is inconsistent and a sys_file entry linked
from sys_file_reference does not exist anymore, an exception
is shown in the "page" module in the backend:

> No file found for given UID.

We will now gracefully show the reference as "broken" and let
the action up to the editor. Works for temporarily unavailable
storages too.

Releases: master, 6.2
Resolves: #58013
Change-Id: Ibce9bb3b60e9294aa1902b95c1b99337d3a8e2e6
Reviewed-on: http://review.typo3.org/36489
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>