[BUGFIX] Show file instead of file_reference info
The irre elements of FAL items have an info icon which shows
some details info about the file_reference. But this is not
the info an editor is looking for he wants the file info.

This patch changes the link behind the info icon for FAL items
so it shows the file info.

Resolves: #57302
Resolves: #57832
Releases: 6.2
Change-Id: I3c4b867ceb5bc05e1cb62525450102784793fadc
Reviewed-on: https://review.typo3.org/29651
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Sebastian Fischer
Tested-by: Nico de Haen
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters