[TASK] Remove unused EM_CONF variables
The ext_emconf.php are created when e.g. an extension
is downloaded from TER. Based on the data from TER the
extension information is built, but not needed anymore.

There are some values that are not in use in the current
6.2 system anymore, and just make ext_emconf.php
more complicated, as the current EM only runs with TER2
anyway.

The patch cleans up values when writing ext_emconf.php
for an extension. As soon as only 6.2+ is supported, the
TER could remove the values to be shipped in the
xml.gz files, thus saving resources (not much but I'm in
favor of a clean system).

Releases: 6.2,6.3
Resolves: #58365
Change-Id: Ic15247f57b557971e9a7c608e4dadd2080e976a3
Reviewed-on: https://review.typo3.org/31399
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn