[BUGFIX] IMAGE param stdWrap-function not working
Images in TypoScript are rendered using method cImage() of
the ContentObjectRender. This method has the possibility to
modify the HTML parameters to be set.

However, the parameters are not recognized, if those are only
rendered using stdWrap functions, e.g. using params.dataWrap

This change moves method stdWrapValue() from the FILES cObject
to the parent ContentObjectRenderer and delegates the call to
the new location of that method.
Besides that, a superfluous space has been removed from the
template markers definition.

Resolves: #58420
Releases: 6.2
Change-Id: I56229ef5aa5133205114c6ee945a68983ef3f52e
Reviewed-on: https://review.typo3.org/29796
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Jan Runte
Tested-by: Jan Runte
Reviewed-by: Wouter Wolters
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader