[BUGFIX] Clear all cache in install tool throws exception
reinitializeClassLoaderAndCachesAndPackageManagement
not only reloads the class loader and package management
but also reloads LocalConfiguratzion and
AdditionalConfiguration. This can lead to errors if
AdditionalConfiguration contains require_once calls.
The method is only used by ClearCacheService which
does not need the configuration reload.

Solution is to remove the method and introduce a new
one to unregister only the class loader. The other
bootstrap methods are now called directly in clear
cache of install tool.

Resolves: #58337
Releases: 6.2
Change-Id: I59cec5f3fc7dacbf83854c0559b4a8b3f143707d
Reviewed-on: https://review.typo3.org/29811
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Jan Runte
Tested-by: Jan Runte
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring