[BUGFIX] Remove non accessible pages from browse menu
This patch enables the browse menu to skip inaccessible items
for prev and next links.

It makes use of the filterMenuPages method, which is already
applied to any kind of menu. Since it has not been applied
to the sections of the browse menu type, which are actually
menus themselves, inaccessible pages made browsing unusable.

This applies to missing translations as well.

Change-Id: I8779731139785dea72b078567912c96e3ad5c7ca
Resolves: #58792
Releases: 6.3, 6.2
Reviewed-on: http://review.typo3.org/32333
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>