[BUGFIX] Fix import for fields of type input with link wizard
Fields of type import and link wizard could also contain a path to
a local file in addition to a direct file reference (file:42). The
refindex correctly analyse this relation and saves a db relation to 
the uid of the sys_file records.

On imports this is not correct resolved back, but a value
"sys_file_{uid}" is saved.

In addition this patch fixes also a redundant inclusion of files
stored in fileadmin for such a field type, because of having a
sys_file db relation and a softreference.

Resolves: #58843
Releases: master, 6.2
Change-Id: I32bb0e82212295f71c22dfa4935d7d940eb3c95e
Reviewed-on: http://review.typo3.org/30171
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>