[BUGFIX] Fix import for fields of type input with link wizard
Fields of type import and link wizard could also contain a path to
a local file in addition to a direct file reference (file:42). The
refindex correctly analyse this relation and saves a db relation to
the uid of the sys_file records.

On imports this is not correct resolved back, but a value
"sys_file_{uid}" is saved.

In addition this patch fixes also a redundant inclusion of files
stored in fileadmin for such a field type, because of having a
sys_file db relation and a softreference.

Change-Id: I1399006b44f6a9365a92e42911cc3daaa8a81d89
Resolves: #58843
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/36384
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>