[BUGFIX] Missing and invalid warnings on referenced files
The file list module in the backend offers the possibility to
delete files. An interesting use case is the deletion of files
that still have references to other entities. However, currently
proper warnings are only available for sys_file reference.

a) before deleting a file, a popup should be shown
   this does not work since the API is called with
   the identifier instead of a proper UID value

b) after calling the API for deleting a file, references
   should be pointed out
   references are checked, however the description on the
   file usage is missing due to hardcoded "sys_file_reference"
   triggers - thus, no other references are printed out

Resolves: #58913
Releases: master, 6.2, 6.1
Change-Id: Id80361427abc17bda3461c2d00750c4790798c88
Reviewed-on: http://review.typo3.org/33495
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>