[BUGFIX] Avoid superfluous IRRE child record duplication
If copying a page, all records on that page will be copied to
the accordant destination page. IRRE parent-child structures
are cloned along the way as well. However, if a table (that is
defined a IRRE child) is processed before the accordant parent
record, the parent itself will duplicate its children again.
This behaviour leads to superfluous duplicates and is wrong.

A check in DataHandler::copyRecord_procBasedOnFieldType() now
ensures that records are only copied once during the accordant
DataHander copy process.

Resolves: #44795
Releases: 6.2, 6.1
Change-Id: Iebc22529a98b73cc55396de0718fee9aa877b5bb
Reviewed-on: http://review.typo3.org/30469
Reviewed-by: Felix Jacobi <flex.jacobi+t3@gmail.com>
Tested-by: Felix Jacobi <flex.jacobi+t3@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>