[BUGFIX] Preview of moved workspace elements on source page
If e.g. a content element gets moved to a different page, the
workspace preview link opens the source page (which does not
contain the element anymore, because it was moved).

Since the elements are shown on the correct page in the
workspace backend module, using the target page in the preview
link for the frontend is just obvious.

Resolves: #59143
Releases: 6.2
Change-Id: Ib84e65766e825cc26a66cbda9ffa8132c3a3ec18
Reviewed-on: https://review.typo3.org/30474
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader