[TASK] Reduce complexity in frontend functional tests
Testing the frontend behavior in functional tests creates
currently an own PHP sub-request to execute TypoScript in
frontent context.
Several hooks collect called cObjects and try to find the
correct nesting levels. Albeit this is working for the
current tests, it is way too complex for extending the
tests for further scenarios like for FAL and Extbase.

The hook magic is resolved and explicit render functions,
called via TypoScript, collect the data an output aggregated
JSON data in the end. Besides that the PHP sub-request
execution now re-uses the PHPUnit utilities to determine
and run the accordant executables.

Resolves: #59487
Releases: 6.2
Change-Id: Ic9eeadee5471a3e0e40d2dffb69b38dec16a0aa1
Reviewed-on: https://review.typo3.org/30515
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader