[BUGFIX] Fix PHP Warning on executing Database Analyzer without changes
When running the Database Analyzers "Execute" without checking at least
one checkbox you get a PHP Warning "array_merge(): Argument #2 is not
an array" because instead of returning an array with a message only the
message string was returned.

Fix this by returning the correct variable, which is of type array.

Resolves: #59504
Releases: 6.2
Change-Id: Ifd73d617a4a97f07151fe0e6c6971bcd79c9709c
Reviewed-on: https://review.typo3.org/30670
Reviewed-by: Markus Klein
Tested-by: Markus Klein