[BUGFIX] Make Extbase respect sys_language_mode=strict
The Extbase query parser currently doesn't respect strict language mode
because it doesn't consider records that are only present in a localized
language version, but not in the default language.

As a consequence, records only present in a localized language version
are not fetched at all while for counting/paging the count of the default
language version is used.

This change adds a check for strict language mode and additional query
parts to solve this issue.

Releases: 6.3, 6.2
Resolves: #47183
Change-Id: Ifb0ff6dda648fb4a686992ee6f7c00419a3ee934
Reviewed-on: https://review.typo3.org/31061
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer