[BUGFIX] language condition does not match exact string
The documentation for conditions of type "language" states: "The values
must be a straight match with the value of getenv('HTTP_ACCEPT_LANGUAGE')
from PHP." (or you can use wildcards). Exact matches are actually
not working - and the unit test showing the problem was skipped.

Patch adds the feature as documented and adjusts the unit test.

Change-Id: I5691fff41c889476d8149d6f92c26731ad7bb1fd
Resolves: #59707
Releases: 6.3
Reviewed-on: https://review.typo3.org/30920
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn