[FEATURE] Add option to disable automatic installation of extensions
The extension manager currently downloads and installs an extension
from TER or after file upload directly. Some people want to do some
audits or modifications before installing an extension, therefore there
should be an option to disable the automatic installation process and
just download an extension like it was default in older extension managers.

Like suggested in the related UX/forge issue this patch implements
a simple switch in the configuration of the extension manager itself.

Resolves: #50501
Related: #50582
Releases: master
Change-Id: I538306fd9a067abebe0ffd36704f95eb52c3c454
Reviewed-on: http://review.typo3.org/30972
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>