[TASK] Re-work/simplify copyright header in PHP files - Part 1
This patch replaces the copyright/license header in PHP files with a
new, simplified one. The new header does not show the year figure, nor
an author name, and refers to the LICENSE.txt file for the full
copyright information. License is: GPL2 or any later version.

This is a multi-part commit due to the huge number of changed files.
See issue #59777 for further details.

Resolves: #59777
Releases: 6.3, 6.2
Change-Id: Idf5ef5a0da53dd052af54d42a115e951bc9d0e91
Reviewed-on: https://review.typo3.org/31048
Reviewed-by: Markus Klein
Tested-by: Markus Klein