[TASK] Re-work/simplify copyright header in PHP files - Part 5
This patch replaces the copyright/license header in PHP files with a
new, simplified one. The new header does not show the year figure, nor
an author name, and refers to the LICENSE.txt file for the full
copyright information. License is: GPL2 or any later version.

This is a multi-part commit due to the huge number of changed files.
See issue #59781 for further details.

Resolves: #59781
Releases: 6.3, 6.2
Change-Id: I7429b8e9c507a4833215750268d1975402679828
Reviewed-on: https://review.typo3.org/31026
Reviewed-by: Krzysztof Adamczyk
Tested-by: Krzysztof Adamczyk
Reviewed-by: Markus Klein
Tested-by: Markus Klein