[TASK] Re-work/simplify copyright header in PHP files - Part 6
This patch replaces the copyright/license header in PHP files with a
new, simplified one. The new header does not show the year figure, nor
an author name, and refers to the LICENSE.txt file for the full
copyright information. License is: GPL2 or any later version.

This is a multi-part commit due to the huge number of changed files.
See issue #59782 for further details.

Resolves: #59782
Releases: 6.3, 6.2
Change-Id: Idcd11219eb132c8690cfde256950980e9423bafd
Reviewed-on: https://review.typo3.org/31052
Reviewed-by: Markus Klein
Tested-by: Markus Klein