[TASK] Re-work/simplify copyright header in PHP files - Part 7
This patch replaces the copyright/license header in PHP files with a
new, simplified one. The new header does not show the year figure, nor
an author name, and refers to the LICENSE.txt file for the full
copyright information. License is: GPL2 or any later version.

This is a multi-part commit due to the huge number of changed files.
See issue #59783 for further details.

Resolves: #59783
Releases: 6.3, 6.2
Change-Id: I0e2f68990217f7442abe5b940fd769250c37aec0
Reviewed-on: https://review.typo3.org/31053
Reviewed-by: Markus Klein
Tested-by: Markus Klein