[TASK] Re-work/simplify copyright header in PHP files - Part 8
This patch replaces the copyright/license header in PHP files with a
new, simplified one. The new header does not show the year figure, nor
an author name, and refers to the LICENSE.txt file for the full
copyright information. License is: GPL2 or any later version.

This is a multi-part commit due to the huge number of changed files.
See issue #59784 for further details.

Resolves: #59784
Releases: 6.3, 6.2
Change-Id: I1eacf9bf4537776e850d091d9d50bc72f8c2b20d
Reviewed-on: https://review.typo3.org/31054
Reviewed-by: Markus Klein
Tested-by: Markus Klein