[BUGFIX] Re-introduce read-only file mounts
This re-introduces the read-only file mounts that were silently dropped
in TYPO3 6.0.

File mounts can be added by User TSconfig the same way they were added
in TYPO3 4.x. But since FAL added Storages, the storage needs to be
configurable.

options.folderTree.altElementBrowserMountPoints = 3:/test

This adds the folder "test" of storage 3 as read-only mount in the File
List and the Element Browser. The old syntax is still supported:

options.folderTree.altElementBrowserMountPoints = documents

If no storage is configured, it is assumed that the folder is in the
default storage.

Resolves: #49391
Resolves: #57979
Documentation: #59648
Releases: 6.3, 6.2
Change-Id: I648c8ae15d4add98cd55e2f9c27d89d8e49d6152
Reviewed-on: https://review.typo3.org/31496
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel