[BUGFIX] Felogin improperly merges flexforms and TS conf
There are two options to show the forgot password link in FE:
 * TypoScript: showForgotPasswordLink
 * Flexforms: showForgotPassword

The flexform values are merged with the TS values, but as these config
options are named differently, they are not overridden properly.

With #50138 the forgot-form is now only displayed when showForgotPassword
is set, hence the TypoScript configuration is not taken into account.

Since we can neither change the flexform, nor the TypoScript, we need to
merge the values in the flexform-merge code part.

This patch reduces the code to use only the documented TypoScript option.

Resolves: #59821
Releases: 6.3, 6.2
Change-Id: I77c391910561a304649d4c747ec868388be3f1a5
Reviewed-on: https://review.typo3.org/31081
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert