[BUGFIX] Wrong nesting of deleted versioned child records
If deleting an IRRE child reference, the nesting leving in the
workspace module is not shown correctly. The reason for that is,
that the sys_refindex entry points to the live id of the deleted
child record.

This change takes care of correct visualization only, but does not
touch the sys_refindex behavior in this case.

Resolves: #59853
Releases: master, 6.2
Change-Id: I09970d12ba10e528c150275952ec4c9ff8333a0a
Reviewed-on: http://review.typo3.org/32138
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>