[BUGFIX] Failing cascaded delete action of localized elements
Having a record in live workspace with an accordant localization
leads to the fact that actions to the default language also have
an impact to all localizations.

In a workspace, first deleting the localization, then deleting
the record of the default language, ends up in actually having
reverted the previous deletion of the localization.

The solution is to check for delete placeholders in
DataHandler::deleteL10nOverlayRecords().

Resolves: #59113
Releases: master, 6.2
Change-Id: I888c83a58d591f57bc2ea75a49fa603a56a7f722
Reviewed-on: http://review.typo3.org/31142
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>