[TASK] Use stable phpunit 4.1 instead of master
We had to update to phpunit "master" recently due to an issue with
phpunit-mock-objects that was not released yet but fixed in master.
The fix was released with phpunit-mock-objects 2.1.5 today, so we
can switch back to a stable combination of those packages.

Resolves: #59958
Related: #59685
Releases: 6.3, 6.2
Change-Id: I92336adef0291239c7dad606002721c91b2991fa
Reviewed-on: https://review.typo3.org/31179
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
(cherry picked from commit ee2dd7719711ec3786a86d8ded7dc5a085a7d7e9)
Reviewed-on: https://review.typo3.org/31182