[BUGFIX] Superfluous SQL queries on copying records
On copying a page in the backend there are lots of superfluous
SQL queries containing "... WHERE uid=0". The reason for that
is, that DataHandler::recordInfo() is called for each field
which then tries to fetch the database record for the given
uid "NEW...". Casting the new record placeholder to an integer
value ends up with the "uid=0" condition.

In a test with copying a page with some sub-pages, there was a
total of about 1300 SQL queries, 300 of them have been those
superfluous "uid=0" SELECT statements.

Resolves: #58357
Releases: 6.2, 6.3
Change-Id: Ib7006a7ceb52376032df8eaefaa918407718a482
Reviewed-on: https://review.typo3.org/31502
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader