[!!!][BUGFIX] Skip cache hash for URIs to non-cacheable actions
When building an URI for a non-cacheable action, while the current
request also is uncached, we can skip the cache hash for the target URI
to avoid unnecessary page cache entries.

Since this is a change in behavior during link generation, which other
code may rely upon, this is marked as breaking change.

Resolves: #60272
Releases: master
Change-Id: I448c33d23b790de1064eff95d0a940878b0299ac
Reviewed-on: http://review.typo3.org/31594
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>