[BUGFIX] Fix resource exceptions
When rendering the FILES content object,
exceptions due to missing file references and missing collections
are meant to be caught. However they are not, because
InvalidArgumentException and RuntimeException are thrown instead of
ResourceDoesNotExistException

Throw the appropriate exceptions instead and also
fix a wrong check in RecordCollection Repository, where
the return type is checked against NULL while the DB
call returns FALSE when no record is found.

Resolves: #60506
Releases: 6.2, 6.3
Change-Id: I9cea492d814850f75549d280c2de0defc8b91548
Reviewed-on: http://review.typo3.org/31757
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>