[BUGFIX] Fetch related records for existing uids only
Before fetching related records from a foreign table, we must ensure
that there is a uid to be considered, since otherwise we will fetch
any record having a 0 value for that field.

There are method calls in the core, that don't provide a uid, so the
default value 0 is used, i.e. when fetching children of translated
gridelements, causing memory outage when tt_content is > 10000 records.

Resolves: #60494
Releases: 6.3, 6.2
Change-Id: I1105964f98f79074bb37dc1921180b493fac4bc6
Reviewed-on: http://review.typo3.org/31947
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>