[TASK] Prevent Inclusion of MediaWizardProvider* in Base Bootstrap
On every Frontend request, the original media wizard
provider of EXT:cms is loaded, by fetching the
MediaWizardProviderManager and then
an object of MediaWizardProvider is loaded and held
in memory, as this is not needed, but only on pages
where a media is used.

This change moves the instantiation of the providers to
the time when it is needed.

Further modification would be to add the registered
wizards to TYPO3_CONF_VARS[FE][mediaProviders] and
use them directly in the MEDIA object, so the
MediaWizardProviderManager class could be deprecated.

Releases: 6.3
Resolves: #56336
Change-Id: Id466d41fef019449d820b6faec3946ec1c951488
Reviewed-on: http://review.typo3.org/28046
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>