[TASK] Remove TCA definitions for pre FAL fields
There are some overrides in the CMS extension that provide the old
TCA for the image and media fields of tt_content and
pages_language_overlay tables. These apply if the compat version
is set to 6.0 and and FAL upgrade wizard is not yet run.

Those are removed now as we only support migrations from one LTS
to the next LTS.

Resolves: #60594
Releases: master
Change-Id: I40b1b4cbc9404d5f4cac7df27029986a2568f2c8
Reviewed-on: http://review.typo3.org/31841
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>