[!!!][TASK] Streamline error and exception handling
Remove the possibility for extensions to provide
error and exception handling configuration in
ext_localconf.php of an extension. This only worked
anyway only for errors or exceptions at a later point
in the bootstrap. We now by definition make it only possible
in LocalConfiguration.php (or AdditionalConfiguration.php).

In addition make the configuration more reliable and understandable:

* Register error handler even if exception handling is not configured.
* Remove option 2 for displayErrors option
* Configure everything only once and early in the bootstrap.

Releases: master
Resolves: #68131
Change-Id: I204f17510fcb10fa8601162687793910fbcf4464
Reviewed-on: http://review.typo3.org/31844
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>