[BUGFIX] Refine the error message of checkMaximumFileUploadSize
The Install Tool check the PHP configuration for the maximum file
upload size within the checkMaximumFileUploadSize() method.
This method issues an error message in case the values do not fit.
The error message outputs the values in MB which might give a
wrong indication to the user, as the values seem to be equal
although the check complains they are not.

Provide better output by giving those numbers in KB rather than MB,
which will show diverging numbers correctly then.

Resolves: #60627
Releases: 6.3, 6.2
Change-Id: I5faa4ceb4d7279278182716c31bf6d6dcb851052
Reviewed-on: http://review.typo3.org/31873
Reviewed-by: Christian Hünniger <huenniger@mellowmessage.de>
Tested-by: Christian Hünniger <huenniger@mellowmessage.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>