[BUGFIX] QueryGenerator getTreeList negative ids
When persistence.recursive is configured when using a command controller no
records can be retrieved because the storagePid is negated in the
AbstractConfigurationManager, while claiming all implementations of getTreeList
support negative ids. Using abs() on the id when it is negative fixes this.
Also see the ContentObjectRenderer implementation of getTreeList

Change-Id: I7732290126a83a79204b8224d52a3cd878360e64
Resolves: #60697
Releases: 6.3, 6.2
Reviewed-on: http://review.typo3.org/32331
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>