[BUGFIX] Overriding TCA results in Exception
This problem only occours when using "tree" as
display mode for TCA-type "select".
FlexForm-API brings a special configuration to add,
edit and remove items from a selectbox, but in
case of a select-tree it does not make sense. A
completely different API was called to render this tree.
This patch prevents modifying the select-tree.

Releases: 6.3, 6.2
Resolves: #54490
Change-Id: I3ad3c5c608d7968ba90117aa048014b00cfd44af
Reviewed-on: http://review.typo3.org/32643
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>