[BUGFIX] Make sure all new files are indexed
If the FAL indexer scheduler task finds a index record with
same content hash as a new found file the file isn't added to the
index if it isn't marked as missing.

This patch changes the indexer so that new files with same
content hash are added when no index record is found for a missing
file with same content hash.

Resolves: #60764
Releases: 6.3, 6.2
Change-Id: Ifb7f9e6f48ba327c958a7aa0e0b71f0575cabdf7
Reviewed-on: http://review.typo3.org/32269
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>