[BUGFIX] Keep phpunit 4.1 branch
composer.json is configured to fetch latest phpunit 4.x branch.
Since 4.2 was just released stable, it will be used now.

phpunit 4.2 now relies on Ocramius/Instantiator to create mocks
that do not call a constructor during instantiation. This project
was created to cope with a recent security fix in PHP unserialize()
that rendered the old way of creating those mocks invalid.

The instantiator project however has issues with PHP versions
5.3 and 5.6 that lead to some failing tests and need further
investigation.

To not run into this during our next merges, the patch nails
phpunit to 4.1.x for now.

Change-Id: I5feb8346e0713fefe6c8fa899c76d3e040dd1973
Resolves: #60923
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/32116
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>