[BUGFIX] Avoid DB error on writing to log
The writeLog() function of the ErrorHandler tries to log the IP that
caused the error. On CLI there is no IP hence a NULL value is being
tried to insert into the DB, where the DB field is NOT NULL.

Fix this by casting the IP to a string.

Resolves: #61051
Releases: 6.3, 6.2
Change-Id: Id46e97468bf825307258e907811d5f2c77c62161
Reviewed-on: http://review.typo3.org/32247
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Daniel Hürtgen <daniel@higidi.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>