[BUGFIX] Fix broken form wizard with MSIE
When "Behavior / Confirmation page" checkbox is checked in a
form content, MSIE fails to show the fields in right panel.
Hovering right panel's fields shows an empty blue box and an empty tooltip.
This bug only occurs when reloading forms with the checkbox enabled
or when enabling the checkbox.

This patch fixes a JS function that sets the value and then refreshes
the right panel when clicking on one of this kind of elements.

Resolves: #59305
Releases: master, 6.2
Change-Id: I7e4cf72cd06a92a4dbdc929b214c79ca6dcd52b2
Reviewed-on: http://review.typo3.org/34032
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>