[BUGFIX] Make TCA property "exclusiveKeys" work with single key
It is necessary that "exclusiveValues" is a string in the JavaScript
method FormEngine.setSelectOptionFromExternalSource() because
in JavaScript strings (but not numbers) have a method match() which
is used in FormEngine.setSelectOptionFromExternalSource().

"exclusiveValues" representing a single key (e.g. -1) are of type
"Number" - in this case an JavaScript error is thrown when the
script calls match() on it.

This commit ensures that "exclusiveValues" is always casted to
String.

Resolves: #61276
Releases: 6.3, 6.2
Change-Id: I049a5e6d11b3b27be901baa1e00429ab200b9a04
Reviewed-on: http://review.typo3.org/32600
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>