[BUGFIX] Fix loading of registry entries for namespace
When calling get(), all entries for the given namespace are loaded only
if no value for this namespace has been set() before in the same run.

This patch fix this by tracking the loading of namespaces and check the
state before set() or get() is processed. In case the namespace was not
loaded before, the loadEntriesByNamespace() is called to prevent
overwrite of values.

Resolves: #53182
Releases: 6.3, 6.2
Change-Id: Ib09e5892e35cef998413d2f0d8039ed49a7e7226
Reviewed-on: http://review.typo3.org/32619
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>