[TASK] Introduce .editorconfig file
A lot of IDEs support a .editorconfig file, this file contains
informations for the IDE how to handle and format files.
This little file can be used for any kind of file. If the IDE
support this file, the IDE configure it selfs to use the settings
e.g. to use TAB instead of SPACES.

Resolves: #61621
Releases: 6.2, 6.3
Change-Id: I76fb7901c0cf07a5b848523a0c0e46a6f705db4d
Reviewed-on: http://review.typo3.org/32794
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>