[BUGFIX] Follow up: Remove PHP warnings when no icon is found
With https://review.typo3.org/32616/ a check with file_exists() was
introduced to prevent warnings on accessing non-image files. The check
returns true for files and folders. Now warnings are thrown, if an image
icon links to a folder. This patch replaces file_exists() with is_file()
to prevent further warnings.

Resolves: #61640
Releases: 6.3, 6.2
Change-Id: If2f7db44588d7fb9f90d05fb3192653002fe0202
Reviewed-on: http://review.typo3.org/32806
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>