[FEATURE] Migrate open-docs counter to .badge
During streamline of the backend we make use of the Bootstrap .badge
component. The Badge can be attached to every toolbar / menu item to
show a notification / counter.

An input field is the wrong component: no input is possible.
Also CSS can be removed for the former inline input.

Resolves: #61675
Releases: master
Change-Id: Ie0c0d15b84d43c54a5f72ad2e564c2990832612e
Reviewed-on: http://review.typo3.org/32820
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>