[BUGFIX] Possible race condition when creating directory
During various actions in the backend which require a cache directory
in typo3temp/ a race condition can occur while creating the directory.
The race condition is caused by GeneralUtility::createDirectoryPath.

There is a check whether the directory to be created already exists.
Later the directory is created using mkdir(). If in the meantime
another process creates the directory the exception will get thrown.

Resolves: #47269
Releases: master, 6.2
Change-Id: Iff46addd9ad6d738d1760f82a436dec9840c331d
Reviewed-on: http://review.typo3.org/32881
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>