[BUGFIX] Do not check file extension on local file
If a file is uploaded and file extension deny pattern is set,
we should not test this temporary file for extension permissions.
Also on adding a file into a storage do not need to test the
extension of the local file, the extension of the target file is
enough.

Resolves: #58367
Releases: master, 6.2
Change-Id: I95b1685746150e2b03c56a5680785ae0b2388cd9
Reviewed-on: http://review.typo3.org/32930
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Andreas Kiessling <andreas.kiessling@web.de>
Tested-by: Andreas Kiessling <andreas.kiessling@web.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>