[TASK] Remove styling code in FormEngine/TCA
FormEngine uses a combination
of classes and styles, which is very
complicated and not usable. A new way
of handling the styling is needed.
All old methods are deprecated and not in use
anymore.

At a later step, proper classes can be introduced
and styling can be done in LESS cleanly
without any PHP magic.

Releases: master
Resolves: #61890
Change-Id: Ie72226e1c6d795b3d8b53220fbcbee9e059935a6
Reviewed-on: http://review.typo3.org/32975
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>