[TASK] Deprecate unused/non-unified DocumentTemplate code
There are certain parts in DocumentTemplate - one of the main rendering
classes for the Backend - that are not in use anymore or can be
accessed via the pageRenderer, which DocTemplate uses under the hood.

Additionally, certain options, like rendering XHTML or HTML3 in
the backend, are not needed anymore, as the whole code is targeted
to be HTML5 compatible.

The variables

* $inDocStyles
* $inDocStyles_TBEstyle

are unified within the main inDocStylesArray.

JScodeLibArray is not in use anymore and deprecated as well.
The method menuTable is not used anymore and also deprecated.

Releases: master
Resolves: #62038
Change-Id: I136fbd4e58b63d44d048e1b355413bff044be771
Reviewed-on: http://review.typo3.org/33167
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>