[BUGFIX] Show empty icon in checkbox'ed select field
When a select field has config option renderMode=checkbox, icons
are rendered between checkbox and the label. If no icon is defined,
TYPO3 ought to show an empty icon.
This is broken and a red question mark appears.
This patch fixes the problem so that the empty icon is rendered again.

Change-Id: I2c54695463d98bd7b8bfa66642bd861e7a2ca7a8
Resolves: #62116
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/33266
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>