[BUGFIX] RTE Spellcheck:default ISO code dictionary should be lowercase
When no default dictionary is configured for the ISO code of the
language of current content, the ISO code itself is taken as the name
of the Aspell dictionary. This should be lowercase.

Releases: master, 6.2
Resolves: #62128
Change-Id: I346469b0c1cd8a44227576b0d17cf231c205f961
Reviewed-on: http://review.typo3.org/33228
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>
(cherry picked from commit f6d66f21fccee997e0d5f78d42703d23177c4c88)
Reviewed-on: http://review.typo3.org/33229