[BUGFIX] linkvalidator: Fatal error: Call to a member function getUrl()
The external link checker tries to provide detailed information in case
an URL is not accessible. To do this, it analyses the stack trace
provided by the exception.

Replace the stacktrace parsing with a more reliable method by checking
the last event. In case of a loop, the last event contains the response
object in the 'data' key.

Resolves: #62266
Releases: master,6.2
Change-Id: I51aaf3d6af85877541e920afaee76e6cac368e7b
Reviewed-on: http://review.typo3.org/35471
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>