[BUGFIX] Fix suggest wizard for new CE in flexforms
When having a suggest wizard inside a flexform on a 
new content element edit form, we have the problem that
we can't load the wizard's config in the AJAX request due
to missing record information. We don't know what flexform
definition to load, as we don't know what ctype is requested.

Two possible solutions exist:
 - Generate the blank record again in the AJAX request, which
   means we have to copy lots of code from the DataPreProcessor.
   Moreover we would still need to pass the defVal GET parameter
   of the EditDocumentController to the AJAX request handler
   somehow, since we need that information in the DataPreProcessor
   to generate the correct dummy record.
 - Push the record to the AJAX request as serialized array

I chose the second implementation.

Change-Id: If7f8abe66554a3164bbfe98ad396fdf8448f5569
Resolves: #50549
Releases: master, 6.2, 6.1
Reviewed-on: http://review.typo3.org/33570
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>