[BUGFIX] Handling of umlauts/accents in convertFileLinks in upg. wizard
This change fixes an error in the install tool upgrade wizard that outputs
<<... not found in referenced element (uid: ...). Reference index was
probably out of date.>> on RTE content elements that have file links to files
with umlauts / accents.

The filename part in RTE has apparently been rawurlencoded. The change
puts the rawurlencoded filename in the regex instead of just only
replacing whitespace with %20.

Resolves: #59133
Related: #58898, #59133, #62307
Releases: 6.2
Change-Id: Icbc78484652a6f2871c959cefc608fb30f52bbe8
Reviewed-on: http://review.typo3.org/33976
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>